Cleanup cli init file and introduce main entrypoint
