+++ pyproject.toml
@@ -47,7 +47,7 @@ dev = [
 "Bug Tracker" = "https://github.com/Snowflake-Labs/snowcli/issues"
 
 [project.scripts]
-snow = "snowcli.cli:app"
+snow = "snowcli.__main__:main"
 
 [tool.hatch.version]
 path = "src/snowcli/__about__.py"
