+++ src/snowcli/__main__.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+import sys
+
+from snowcli.cli.app import app
+
+
+if __name__ == "__main__":
+    app()
+
+if getattr(sys, "frozen", False):
+    app(sys.argv[1:])
