+++ src/snowcli/cli/snowpark/cp.py
@@ -1,8 +1,7 @@
 import typer
-from snowcli import config
-from snowcli.cli import DEFAULT_CONTEXT_SETTINGS
+
 from snowcli.cli.common.alias import build_alias
-from snowcli.cli.common.flags import ConnectionOption
+from snowcli.cli.common.flags import ConnectionOption, DEFAULT_CONTEXT_SETTINGS
 from snowcli.snow_connector import connect_to_snowflake
 from snowcli.output.printing import print_db_cursor
