+++ src/snowcli/cli/snowpark/services.py
@@ -1,12 +1,10 @@
 import sys
 
 import typer
-from snowcli import config
-from snowcli.cli import DEFAULT_CONTEXT_SETTINGS
 from typing import TextIO
 from typing_extensions import Annotated
 
-from snowcli.cli.common.flags import ConnectionOption
+from snowcli.cli.common.flags import ConnectionOption, DEFAULT_CONTEXT_SETTINGS
 from snowcli.snow_connector import connect_to_snowflake
 from snowcli.output.printing import print_db_cursor
